Description :-
Moved mongoose schema and connection code to another separate file Post.js.
Better naming of variable.
