+++ app.js
@@ -1,43 +1,9 @@
 var express = require('express');
-var mongoose = require('mongoose');
 var app = express();
 app.use(express.static(__dirname + '/public'));
 app.set('views','./views');
 app.set('view engine', 'ejs');
-
-mongoose.connect('mongodb://localhost/test');
-
-var schema = mongoose.Schema;
-
-var blogSchema = new schema({
-  title : String,
-  content : String,
-  DOC : Date
-});
-
-
-// Create a model
-var Blog = mongoose.model('blog',blogSchema);
-
-var createblogPost = function(blogTitle, blogContent, blogDate) {
-
-  // create document
-
-    var blogInstance  = new Blog({
-      title : blogTitle,
-      content : blogContent,
-      DOC : blogDate
-    });
-
-    blogInstance.save(function(err,savedDate){
-      if (err) {
-        console.log('Error is :: '+err);
-        return null;
-      } else {
-        console.log('Blog details saved in DB:: '+ blogInstance);
-      }
-    });
-};
+var Blog = require('./public/javascripts/Post.js');
 
 
 // Writing a middleware function that gives date.
@@ -53,24 +19,11 @@ app.get('/', getDate, function(req,res){
     if (err) {
       console.log('Error is :: '+err);
     } else {
-      res.render('index',{name:data});
+      res.render('index',{blogData:data});
       res.end();
     }
   });
 });
-
-// retrieve all blog post data from mongo db and show them to the client
-app.get('/showAll',getDate, function(req,res){
-  Blog.find({}, function(err, data){
-    if (err) {
-      console.log('Error is :: '+err);
-    } else {
-      res.send(data);
-      res.end();
-    }
-  });
-});
-
 
 app.get('/new',getDate, function(req,res){
   res.render('new');
@@ -80,6 +33,27 @@ app.get('/new',getDate, function(req,res){
 /**
  *  Creats a blog content with title and save that to DB.
  */
+var createblogPost = function(blogTitle, blogContent, blogDate) {
+
+  // create document
+
+  var blogInstance  = new Blog({
+    title : blogTitle,
+    content : blogContent,
+    DOC : blogDate
+  });
+
+  blogInstance.save(function(err,savedDate){
+    if (err) {
+      console.log('Error is :: '+err);
+      return null;
+    } else {
+      console.log('Blog details saved in DB:: '+ blogInstance);
+    }
+  });
+};
+
+
 
 app.get('/createBlog',getDate, function(req,res){
   console.log("Date is "+req.date);
@@ -88,6 +62,7 @@ app.get('/createBlog',getDate, function(req,res){
       blogDate = req.date;
   console.log('Blog Date is ::: '+blogDate);
 
+
   createblogPost(blogTitle, blogContent, blogDate);
   res.redirect('/');
   res.end();
@@ -95,3 +70,6 @@ app.get('/createBlog',getDate, function(req,res){
 
 
 app.listen('8080');
+
+
+//To do :- Move all code related with moongoose to another File namely , Post.js .
