+++ views/index.ejs
@@ -36,7 +36,7 @@
 
     <!-- Show the response form the DB -->
 
-    <% name.forEach(function(value,key){ %>
+    <% blogData.forEach(function(value,key){ %>
         <div data-blog = 'detail'>
             <h1> <%= value.title %> </h1>
             <p class="navbar-text"> <%= value.content %> </p>
